Removed Duplicate Code
